libhns: Bugfix for allocating and freeing sq db buffer
commit 64a169f9aac276dfa98c6f9d5c62cd2cb11a6f67 upstream.

When flush cqe, it needs to allocate sq db buffer for storing sq
pi pointer and using when modifying qp to error state. It needs
to free sq db buffer when destroying qp. These operations will
be triggered, depending on whether or not there is sq. However,
when sq sge number is zero, the sq maybe exist.

Fixes: e1726e934574("libhns: Support flush cqe for hip08 in user space")
Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: Xinliang Liu <z.liuxinliang@hisilicon.com>