libhns: Bugfix for computing valid sge
commit 3d31230ea85a10c5bf3219e333635c3952972c72 upstream.

When user posts a work request with multiple sges, and there is
invalid sge, we need to fill the first and the second valid sges
into wqe. As a result, it needs to use the valiable j to record
the number of filling the valid sge into wqe. when the value of j
is greater than two, it should jump out of the loop.

Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: Xinliang Liu <z.liuxinliang@hisilicon.com>